refactor: moved schedule tools from device tools
